Add mini-batch feature and major refactor.
This commit enables the use of mini-batches in the optimization of the objective function.
The calculation of the loss function and its gradient has also been separated to improve the efficiency of the optimisation algorithms.
Several important functions for the estimation of the KLR model have been refactored.
Finally, several bugs in the mathematical formulae have been corrected.

Note, this commit may break compatibility with previous versions of GKLR.